TEST : BusinessApiTest - 기존의 ControllerTest를 ServiceTest로 변경
